Add cast to suppress sign compare warning

+++ testdir/icv_dim.c
@@ -95,7 +95,7 @@ int main(int argc, char **argv)
    miicv_inqdbl(icv, MI_ICV_NORM_MIN, &dvalue);
    printf(", min = %g\n", dvalue);
    miicv_put(icv, coord, count, ivalue);
-   for (i=0; i< sizeof(ivalue)/sizeof(ivalue[0]); i++)
+   for (i=0; i< (long)(sizeof(ivalue)/sizeof(ivalue[0])); i++)
       ivalue[i] = 0;
    miicv_get(icv, coord, count, ivalue);
    for (i=0; i<3; i++) {
